Fix faulty label when pointing to the end
The labels pointing to the 'last' operation were going over the end of
the ops array which would give an arbitrary value as an address for the
computation of the jump. Now the overflow is tested and in this case the
address of the op is replaced with "prog_size" which is the end of the
program.